Criando DML de Beneficiario e DAL do mesmo, alterando suas procedures para o equivalente

+++ FI.AtividadeEntrevista/DML/Beneficiario.cs
@@ -2,9 +2,9 @@
 {
     public class Beneficiario
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string CPF { get; set; }
         public string Nome { get; set; }
-        public int IdCliente { get; set; }
+        public long IdCliente { get; set; }
     }
 }
